SplitContentView, added documentation and moved `GetTabPageInfos` to interface
